Fixed #20180511_142: allow decimal as amount

+++ app/laravel/database/migrations/2017_06_27_184103_add_portions_entries.php
@@ -17,7 +17,7 @@ class AddPortionsEntries extends Migration
             $table->integer('portion_id')->nullable()->after('product_id')->unsigned()->index();
             $table->foreign('portion_id')->references('id')->on('portions')->onDelete('cascade');
 
-            $table->integer('amount')->nullable()->after('portion_id');
+            $table->decimal('amount', 8, 2)->nullable()->after('portion_id');
         });
     }
